pyproject.toml
[project]
name = "albert-api"
version = "1.0.0"
description = "Albert API projects"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "openai==1.43.0",
]

[project.optional-dependencies]
ui = [
    "requests==2.32.3",
    "streamlit_local_storage==0.0.23",
    "streamlit==1.38.0",
]
app = [
    "boto3==1.34.135",
    "botocore==1.34.135",
    "langchain==0.2.15",
    "langchain-community==0.2.15",
    "qdrant-client==1.10.1",
    "redis==5.0.7",
    "uvicorn==0.30.1",
    "fastapi==0.111.0",
    "docx==0.2.4",
    "pyyaml==6.0.1",
    "python-docx==1.1.2",
    "unstructured==0.14.9",
    "python-magic==0.4.27",
    "grist-api==0.1.0",
    "pdfminer.six==20240706",
    "beautifulsoup4==4.12.3",
    "duckduckgo-search==6.2.13",
    "numpy==2.1.1",
]
dev = [
    "ruff==0.6.5",
    "pre-commit==3.6.2",
]
test = [
    "pytest==8.2.2",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 150

[tool.ruff.lint]
ignore = ["F403", "F841"] # import * and never used variables

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["config", "utils", "app"]
forced-separate = ["tests"]

[tool.ruff.lint.isort.sections]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
